data-platform/share-layer/test/loader/requirements.txt
minio>=7.2.0
python-dotenv>=1.0.0
